Externalize creation of Chat collection view layout models
